Fix type definitions
Added style prop to BitmapText type (styleis used in constructor).
Fixed type definitions to work with both PixiJS v4 and v5.
Fixed type definitions for properties of type any being incorrectly replaced with PointLike.
Fixed type definitions for built-in components to be able to use refs.
Fixed type definitions for AppContext which was exported as type, not as value.